refactor: split lecture watch and course create bundles
