+++ docs/dev-logs/2026-06-06-frontend-common-ui-split-bundle-v6.md
@@ -1,0 +1,14 @@
+# Frontend Common UI Split Bundle v6
+
+## Summary
+- Split `LectureWatchPageSections` into hero and main sections.
+- Split `CourseCreatePageSections` into hero, tabs, form, and studio sections.
+- Split `AdminDashboardPage` into reusable overview/sidebar sections.
+
+## Verification
+- `npm run build:frontend`
+- `npm run verify`
+
+## Notes
+- Frontend build passed after the split.
+- Backend test suite passed unchanged at 74 tests.
